+++ apps/indexer/src/modules/swaps/eos-transfers.ts
@@ -1,17 +1,17 @@
 import { smartsaleEnv } from 'smartsale-env'
 
-// required by dfuse/client
-import { issueTokens } from './evm-tranfers'
 import { stringify } from 'viem/utils'
 import { createFirehoseSubscription } from '~/modules/swaps/dfuse-client'
+import { issueTokens } from './cred-issuer'
 
+// https://docs.dfuse.eosnation.io/platform/public-apis/search-query-language/
+// https://docs.dfuse.eosnation.io/eosio/public-apis/reference/search/terms/
+// receiver: means the account with code that has executed the action.
 export async function listenToEosTransfers(env: 'test' | 'prod' = 'test') {
   const usdt = smartsaleEnv[env].usdt.find((t) => (t.chainType = 'antelope'))?.address
   const bank = smartsaleEnv[env].bitcash.bank
   const launchpad = smartsaleEnv[env].smartsale.bk
-  // https://docs.dfuse.eosnation.io/platform/public-apis/search-query-language/
-  // https://docs.dfuse.eosnation.io/eosio/public-apis/reference/search/terms/
-  // receiver: means the account with code that has executed the action.
+
   const usdtDeposits = await createFirehoseSubscription(
     `receiver:${bank} action:stbtransfer data.to:${launchpad}`,
   )
