+++ apps/indexer/src/modules/swaps/evm-tranfers.ts
@@ -1,21 +1,12 @@
-import { EVMTokenContractData, SepoliaUSDT, TestnetUSDCred, TestnetUSDT } from 'smartsale-contracts'
+import { EVMTokenContractData, SepoliaUSDT, TestnetUSDT } from 'smartsale-contracts'
 import { runPromisesInSeries } from '~/utils'
 
-import {
-  Address,
-  Log,
-  PublicClient,
-  createPublicClient,
-  createWalletClient,
-  http,
-  parseAbiItem,
-  stringify,
-} from 'viem'
+import { Address, Log, PublicClient, createPublicClient, http, parseAbiItem, stringify } from 'viem'
 import { ApprovalEvent, TransferEvent } from '~/types'
 import { db } from 'smartsale-db'
 import { sepolia } from 'viem/chains'
-import { eosEvmTestnet, smartsaleChains } from 'smartsale-env'
-import { appenv } from '~/config'
+import { smartsaleChains } from 'smartsale-env'
+import { issueTokens } from './cred-issuer'
 
 const tokens: EVMTokenContractData[] = [SepoliaUSDT, TestnetUSDT]
 
@@ -76,7 +67,7 @@ async function processLogs(logs: Log[], delay = 0) {
       if (!(eventName in eventHandlers)) return null
       return async () => {
         try {
-          eventHandlers[eventName](log)
+          eventHandlers[eventName] && eventHandlers[eventName](log)
         } catch (error) {
           //TODO: sent sentry reports
           console.error(error)
@@ -90,7 +81,6 @@ async function processLogs(logs: Log[], delay = 0) {
 
 const eventHandlers: { [key: string]: (log: any) => void } = {
   Transfer: handleTransfer,
-  Approval: handleApproval,
 }
 
 async function handleTransfer(log: TransferEvent) {
@@ -127,32 +117,4 @@ async function handleTransfer(log: TransferEvent) {
   })
 
   console.log('tokens issued', { usdcred_trx, trx: log.transactionHash })
-}
-
-function handleApproval(log: ApprovalEvent) {
-  console.log('handleApproval', log)
-}
-
-export async function issueTokens(to: Address, amount: bigint) {
-  console.log('issueTokens', {
-    args: [to, amount],
-  })
-
-  try {
-    const walletClient = createWalletClient({
-      chain: eosEvmTestnet,
-      transport: http(),
-      key: appenv.evm.issuerKey,
-      account: appenv.evm.issuerAccount,
-    })
-    return walletClient.writeContract({
-      address: TestnetUSDCred.address,
-      abi: TestnetUSDCred.abi,
-      functionName: 'issue',
-      args: [to, amount],
-    })
-  } catch (error) {
-    console.log((error as Error).message)
-    return null
-  }
 }
