+++ package.json
@@ -17,6 +17,8 @@
     "eslint": "^8.44.0",
     "prettier": "^3.2.5",
     "prisma": "latest",
+    "ts-node": "^10.9.2",
+    "ts-node-dev": "^2.0.0",
     "turbo": "^1.11.2"
   },
   "workspaces": {
